feature : Added custom hooks.
